Publish kv-cache events
