+++ pkg/common/publisher.go
@@ -17,6 +17,7 @@ limitations under the License.
 package common
 
 import (
+	"bytes"
 	"context"
 	"encoding/binary"
 	"errors"
@@ -24,7 +25,7 @@ import (
 	"sync/atomic"
 
 	zmq "github.com/pebbe/zmq4"
-	"github.com/vmihailenco/msgpack"
+	"github.com/vmihailenco/msgpack/v5"
 	"k8s.io/klog/v2"
 )
 
@@ -62,7 +63,11 @@ func NewPublisher(endpoint string) (*Publisher, error) {
 func (p *Publisher) PublishEvent(ctx context.Context, topic string, batch interface{}) error {
 	logger := klog.FromContext(ctx).V(0)
 
-	payload, err := msgpack.Marshal(batch)
+	// Use an encoder configured for struct as array
+	var payload bytes.Buffer
+	enc := msgpack.NewEncoder(&payload)
+	enc.UseArrayEncodedStructs(true)
+	err := enc.Encode(batch)
 	if err != nil {
 		return fmt.Errorf("failed to marshal event batch: %w", err)
 	}
@@ -73,7 +78,7 @@ func (p *Publisher) PublishEvent(ctx context.Context, topic string, batch interf
 	binary.BigEndian.PutUint64(seqBytes, seq)
 
 	// send topic, sequence, payload
-	if _, err := p.socket.SendMessage(topic, seqBytes, payload); err != nil {
+	if _, err := p.socket.SendMessage(topic, seqBytes, payload.Bytes()); err != nil {
 		return fmt.Errorf("failed to send message to topic %s: %w", topic, err)
 	}
